pyproject.toml
# https://docs.astral.sh/ruff/configuration/
# https://docs.astral.sh/ruff/rules/

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 100
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "YTT",
    "FBT",
    "A",
    "COM",
    "C4",
    "T10",
    "EXE",
    "ICN",
    "PIE",
    "Q",
    "RET",
    "SIM",
    "TID",
    "PTH",
    "TD",
    "FAST",
    "PERF",
    "FURB",
    "RUF",
]
ignore = [
    "COM812",
    "TD003",
    "TD004",
    "RET504",
    "UP006",
    "UP035",
    "N805",
    "FBT002",
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

# Add uv project configuration
[project]
name = "draw_flags"
version = "0.1.0"
description = "Draw a flag, have it recognized"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.33.0",
    "pydantic==2.10.4",
    "requests",
    "tokenizers",
    "numpy==1.26.4",
    "sympy==1.14.0",  # Pin to avoid compilation
    "onnxruntime==1.22.1",  # Pin to avoid sympy compilation
    "pillow",  # For image processing
    "python-multipart",  # For file uploads in FastAPI
    # Note: sentence-transformers is lazy-loaded only for image processing
    "sentence-transformers==3.2.1",  # Required for image encoding (lazy-loaded)
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.1",
    "pytest",
    "pytest-asyncio",
    "black",
    "mypy",
    "pre-commit",
    "sentence-transformers==3.2.1",
    "torch",
    "aiohttp==3.10.11",
    "rich>=13.9.4",
    "tqdm==4.66.5",
    "levenshtein==0.25.1",
    "coloredlogs==15.0.1",
    "onnx>=1.18.0",
    "scikit-learn==1.3.2",
    "transformers>=4.35.0",
    "pillow",
    "cairosvg>=2.7.1",
]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
]
